refactor: 숙소 예약 정보 컴포넌트 분리
