+++ src/app/messages/components/reservation/PaymentInformation.tsx
@@ -1,0 +1,26 @@
+import { IoIosArrowForward } from 'react-icons/io';
+import { TbReceiptFilled } from 'react-icons/tb';
+
+export function PaymentInformation() {
+  return (
+    <div className="flex flex-col gap-5 px-4">
+      <h2 className="text-xl font-semibold">결제 정보</h2>
+      <div>
+        <h3 className="font-bold">결제한 금액</h3>
+        <p>₩207,953.00</p>
+      </div>
+      <div className="-mx-4 px-4 hover:bg-gray-100">
+        <hr />
+        <button className="w-full py-3">
+          <div className="flex items-center justify-between">
+            <div className="flex items-center gap-3">
+              <TbReceiptFilled />
+              <p>영수증 받기</p>
+            </div>
+            <IoIosArrowForward />
+          </div>
+        </button>
+      </div>
+    </div>
+  );
+}
